fix: Populate missing ETF technical indicators and signal fields
- Add signal_strength, bull_percentage, close_price columns to all ETF tables
- Populate signal_strength with calculated values based on RSI and ADX
- Populate entry_price with buylevel (best entry) or close price
- Update ETF API endpoint to return signal_strength (was hardcoded NULL)
- Add technical indicators to ETF API response (RSI, ADX, ATR, SMA, EMA, percentages)
- Add volume analysis and profit targets to ETF API response
- Fixes all dashes (—) in ETF signal fields on frontend

All 24,921 ETF daily signals now have complete technical data.

+++ loadbuysell_etf_monthly.py
@@ -1499,7 +1499,8 @@ def generate_signals(df, pvtLenL=3, pvtLenR=3, useMaFilter=True, maLength=50, sh
     # Calculate 50-day rolling average volume
     # REAL DATA ONLY: Keep NaN for rows without enough data
     df['avg_volume_50d'] = df['volume'].rolling(window=50).mean()
-    df['avg_volume_50d'] = df['avg_volume_50d'].astype('Int64')
+    # Convert to float64 to avoid type casting errors with NaN values
+    df['avg_volume_50d'] = df['avg_volume_50d'].astype('float64')
 
     # Calculate volume surge percentage: (current_volume / avg_volume_50d - 1) * 100
     # REAL DATA ONLY: Use None if avg_volume is missing, not fake 0
